+++ src/main/java/racingcar/util/RandomGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
