+++ src/test/java/racingcar/PositionTest.java
@@ -1,9 +1,9 @@
 package racingcar;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Position;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
