--- src/test/java/racingcar/RacingGameTest.java
@@ -1,15 +1,0 @@
-package racingcar;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-class RacingGameTest {
-    @Test
-    void RacingGame_생성() {
-        RacingGame racingGame = new RacingGame();
-
-        assertThat(racingGame).isInstanceOf(RacingGame.class);
-    }
-
-}
